Калькулятор. Сущности и формы. -fix: Неверный вызов cущностив в форме; -fix: замена DepartmentMedCase на MedCase; -fix: единый тип сущность - формат.

+++ ecom/ecom-riams/ejb/src/main/java/ru/ecom/mis/ejb/domain/calc/Calculations.java
@@ -34,9 +34,9 @@ public class Calculations extends BaseEntity{
 	
 	/** Порядок */
 	@Comment("Порядок")
-	public Integer getOrderus() {return theOrderus;}
-	public void setOrderus(Integer aOrderus) {theOrderus = aOrderus;}
-	private Integer theOrderus;
+	public Long getOrderus() {return theOrderus;}
+	public void setOrderus(Long aOrderus) {theOrderus = aOrderus;}
+	private Long theOrderus;
 	
 	/** Значение */
 	@Comment("Значение")
